+++ cisco_aci/tests/test_fabric.py
@@ -12,8 +12,8 @@ from datadog_checks.cisco_aci.api import SessionWrapper, Api
 from datadog_checks.cisco_aci import CiscoACICheck
 from datadog_checks.utils.containers import hash_mutable
 
-import conftest
-from .common import FIXTURE_LIST_FILE_MAP
+from . import common
+from .mock_sender import mock_send
 
 log = logging.getLogger('test_cisco_aci')
 
@@ -21,7 +21,7 @@ log = logging.getLogger('test_cisco_aci')
 class FakeSess(SessionWrapper):
     """ This mock:
      1. Takes the requested path and replace all special characters to underscore
-     2. Fetch the corresponding hash from FIXTURE_LIST_FILE_MAP
+     2. Fetch the corresponding hash from common.FIXTURE_LIST_FILE_MAP
      3. Returns the corresponding file content
      """
     def make_request(self, path):
@@ -39,12 +39,12 @@ class FakeSess(SessionWrapper):
         mock_path = mock_path.replace(']', '_')
         mock_path = mock_path.replace('|', '_')
         try:
-            mock_path = FIXTURE_LIST_FILE_MAP[mock_path]
+            mock_path = common.FIXTURE_LIST_FILE_MAP[mock_path]
 
-            mock_path = os.path.join(conftest.FABRIC_FIXTURES_DIR, mock_path)
+            mock_path = os.path.join(common.FABRIC_FIXTURES_DIR, mock_path)
             mock_path += '.txt'
 
-            log.info(os.listdir(conftest.FABRIC_FIXTURES_DIR))
+            log.info(os.listdir(common.FABRIC_FIXTURES_DIR))
 
             with open(mock_path, 'r') as f:
                 return json.loads(f.read())
@@ -55,20 +55,20 @@ class FakeSess(SessionWrapper):
 @pytest.fixture
 def session_mock():
     session = Session()
-    setattr(session, 'send', conftest.mock_send)
-    fake_session_wrapper = FakeSess(conftest.ACI_URL, session, 'cookie')
+    setattr(session, 'send', mock_send)
+    fake_session_wrapper = FakeSess(common.ACI_URL, session, 'cookie')
 
     return fake_session_wrapper
 
 
 def test_fabric_end_to_end(aggregator, session_mock):
-    check = CiscoACICheck(conftest.CHECK_NAME, {}, {})
-    api = Api(conftest.ACI_URLS, conftest.USERNAME,
-              password=conftest.PASSWORD, log=check.log, sessions=[session_mock])
+    check = CiscoACICheck(common.CHECK_NAME, {}, {})
+    api = Api(common.ACI_URLS, common.USERNAME,
+              password=common.PASSWORD, log=check.log, sessions=[session_mock])
     api._refresh_sessions = False
-    check._api_cache[hash_mutable(conftest.CONFIG_WITH_TAGS)] = api
+    check._api_cache[hash_mutable(common.CONFIG_WITH_TAGS)] = api
 
-    check.check(conftest.CONFIG_WITH_TAGS)
+    check.check(common.CONFIG_WITH_TAGS)
 
     tags000 = ['cisco', 'project:cisco_aci', 'medium:broadcast', 'snmpTrapSt:enable', 'fabric_pod_id:1']
     tags101 = tags000 + ['node_id:101']